Level editor initial working commit

+++ Assets/Scripts/TileMapGenerator/GameObjectPosition.cs
@@ -1,6 +1,6 @@
 class GameObjectPosition
 {
-    public int x { get; set; }
-    public int y { get; set; }
+    public float x { get; set; }
+    public float y { get; set; }
     public string gameObjectName { get; set; }
 }
